+++ app/scripts/services/dimSettingsService.factory.js
@@ -53,6 +53,9 @@
       itemSize: 44,
       // Which categories or buckets should be collapsed?
       collapsedSections: {},
+      // What settings for farming mode
+      farming: {
+      },
       // Predefined item tags. Maybe eventually allow to add more (also i18n?)
       itemTags: [
         { type: undefined, label: 'Tag Item' },